backend/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
hiredis==2.2.3

# Auth
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# PayPal
paypalrestsdk==1.13.1

# S3/Storage
boto3==1.29.7
botocore==1.32.7

# Email
aiosmtplib==3.0.1
email-validator==2.1.0

# Image processing
Pillow==10.1.0

# HTTP
httpx==0.25.2
aiohttp==3.9.1

# Utils
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
